+++ frontend/src/routes/redirect/index.tsx
@@ -1,6 +1,6 @@
 import { redirect } from "react-router-dom";
 
-import { needRedirect } from "~/utils/redirectNGLocation";
+import { needRedirect } from "~/utils/redirectBlogNGLocation";
 
 /** Check old format URL and redirect if needed */
 export const loader = async () => {
