+++ frontend/src/routes/root/index.tsx
@@ -1,7 +1,7 @@
 import { Layout } from "@edifice-ui/react";
 import { Outlet, redirect } from "react-router-dom";
 
-import { needRedirect } from "~/utils/redirectNGLocation";
+import { needRedirect } from "~/utils/redirectBlogNGLocation";
 
 /** Check old format URL and redirect if needed */
 export const rootLoader = () => async () => {
